added comments, moved into packages
