+++ Week6/Outfit_Generator/data/Clothing.java
@@ -1,11 +1,12 @@
-abstract public class Clothing {  
+package data;
+
+abstract public class Clothing {
 
   private String _name;
   private String _color;
   private String _season;
   private boolean _isFancy;
   private boolean _isPlain;
-  // private boolean _isForWinter;
 
   //// Getters
   public String name() { return _name; }
@@ -13,7 +14,6 @@ abstract public class Clothing {
   public String season() { return _season; }
   public boolean isFancy() { return _isFancy; }
   public boolean isPlain() { return _isPlain; }
-  // public boolean isForWinter() { return _isForWinter; }
 
   //// Setters
   public void setName(String name) { _name = name; }
@@ -21,7 +21,6 @@ abstract public class Clothing {
   public void setSeason(String season) { _season = season; }
   public void setIsFancy(boolean isFancy) { _isFancy = isFancy; }
   public void setIsPlain(boolean isPlain) { _isPlain = isPlain; }
-  // public void setIsForWinter(boolean isForWinter) { _isForWinter = isForWinter; }
 
   //// toString abstract method
   public abstract String toString();
