+++ Week6/Outfit_Generator/data/OutfitGenerator.java
@@ -1,11 +1,15 @@
+package data;
+
 import java.util.*;
 
 public class OutfitGenerator {
 
   // The generator properties
+  // MAX_TRIES is set up to prevent a possible endless loop if
+  // no outfit can be set up with the given articles of Clothing
   private static final int MAX_TRIES = 1000;
-  private String name;
-  private int numPieces;
+  private String name; // Name for the outfit
+  private int numPieces; // Number of pieces the outfit should consist of
   private ArrayList<Clothing> wardrobe; // all Clothing options to build outfit from
 
   //// Constructors
@@ -16,15 +20,18 @@ public class OutfitGenerator {
   public OutfitGenerator(String newName, int newNumbPieces) {
     name = newName;
     numPieces = newNumbPieces;
-    wardrobe = new ArrayList<Clothing>(50);
-  }
-
-  //// Methods
-
-  public void getNewOutfit() {
-
     SampleWardrobe sample = new SampleWardrobe();
     wardrobe = sample.getSampleWardrobe();
+  }
+
+  //// Setters
+  public void setNumPieces(int newNumbPieces) { numPieces = newNumbPieces; }
+  public void setName(String newName) { name = newName; }
+
+  //// Methods
+  // creates a new outfit using random pieces from the wardrobe based on some
+  // matching criteria
+  public void getNewOutfit() {
 
     Clothing item = getRandomItem(); // get the first item from wardrobe
 
@@ -33,9 +40,9 @@ public class OutfitGenerator {
     outfit.add(item);
 
     // to avoid an endless loop in case a good combination cannot be found
-    int safety = 0;
-    // loop while outfit is not complete and safety is less than MAX_TRIES
-    while (outfit.size() < numPieces && safety < MAX_TRIES) {
+    int safetyCounter = 0;
+    // loop while outfit is not complete and safetyCounter is less than MAX_TRIES
+    while (outfit.size() < numPieces && safetyCounter < MAX_TRIES) {
       // get random item
       item = getRandomItem();
       String itemType = item.getClass().getName();
@@ -44,22 +51,26 @@ public class OutfitGenerator {
       if (!outfit.hasType(itemType) && checkMatches(outfit, item)){
         outfit.add(item);
       }
-      safety++;
+      safetyCounter++;
     }
 
-    if(safety < MAX_TRIES) {
+    // if we have a full suit then print all the items
+    // else print a sorry message
+    if(outfit.size() == numPieces) {
       outfit.print();
     } else {
       System.out.println("An outfit could not be found for you. Please try again!");
     }
   }
 
+  // gets a random piece of Clothing from wardrobe
   private Clothing getRandomItem() {
     Random rand = new Random();
     int index = rand.nextInt(wardrobe.size());
     return wardrobe.get(index);
   }
 
+  // one method to call all the other "check" methods
   private boolean checkMatches(Outfit outfit, Clothing item) {
     return fancyMatches(outfit, item) && plainMatches(outfit, item)
        && colorMatches(outfit, item) && seasonMatches(outfit, item);
@@ -70,12 +81,15 @@ public class OutfitGenerator {
     return outfit.isFancy() == garb.isFancy();
   }
 
-  // Return true 50% of the time, so there is some possibility of plain and
-  // colourful articles to mix
+  // Check that outfit and new item both match
+  // Outfit will switch out its Plain attribute to prevent
+  // all plain or all colorful items
   private boolean plainMatches(Outfit outfit, Clothing garb) {
     return outfit.isPlain() == garb.isPlain();
   }
 
+  // reject blue items if the outfit contains green
+  // reject green items if the outfit contains blue
   private boolean colorMatches(Outfit outfit, Clothing garb) {
     String garbColor = garb.color().toLowerCase();
 
@@ -99,7 +113,6 @@ public class OutfitGenerator {
        || (outfitSeason.equals("winter") && garbSeason.equals("not summer"))) {
       return true;
     }
-
     return false;
   }
