+++ Week6/Outfit_Generator/data/Pants.java
@@ -1,11 +1,13 @@
-public class Outerwear extends Clothing {
+package data;
+
+public class Pants extends Clothing {
 
   //// Constructors
-  public Outerwear() {
-    this("Jacket", "Blue", "not summer", false, true);
+  public Pants() {
+    this("Jeans", "Blue", "All", false, true);
   }
 
-  public Outerwear(String name, String color, String season, boolean isFancy, boolean isPlain) {
+  public Pants(String name, String color, String season, boolean isFancy, boolean isPlain) {
     setName(name);
     setColor(color);
     setSeason(season);
@@ -15,7 +17,7 @@ public class Outerwear extends Clothing {
 
   //// toString method needs to over-write the abstract method from Clothing
   public String toString() {
-    return "A " + (isFancy() ? "fancy " : "casual ")
+    return "A pair of " + (isFancy() ? "fancy " : "casual ")
           + (isPlain() ? "plain " : "colorful ")
           + color().toLowerCase() + (season().toLowerCase() == "winter" ? " winter " : " ")
           + name().toLowerCase() + ". ";
