+++ Week6/Outfit_Generator/data/Shirt.java
@@ -1,11 +1,13 @@
-public class Pants extends Clothing {
+package data;
+
+public class Shirt extends Clothing {
 
   //// Constructors
-  public Pants() {
-    this("Jeans", "Blue", "All", false, true);
+  public Shirt() {
+    this("T-shirt", "Green", "all", false, true);
   }
 
-  public Pants(String name, String color, String season, boolean isFancy, boolean isPlain) {
+  public Shirt(String name, String color, String season, boolean isFancy, boolean isPlain) {
     setName(name);
     setColor(color);
     setSeason(season);
@@ -15,7 +17,7 @@ public class Pants extends Clothing {
 
   //// toString method needs to over-write the abstract method from Clothing
   public String toString() {
-    return "A pair of " + (isFancy() ? "fancy " : "casual ")
+    return "A " + (isFancy() ? "fancy " : "casual ")
           + (isPlain() ? "plain " : "colorful ")
           + color().toLowerCase() + (season().toLowerCase() == "winter" ? " winter " : " ")
           + name().toLowerCase() + ". ";
