+++ Week6/Outfit_Generator/data/Shoe.java
@@ -1,3 +1,5 @@
+package data;
+
 public class Shoe extends Clothing {
 
   //// Constructors
